RF: Move output sanitize function to app.utils for more general use
